Add validation

+++ app/routers/user_routers.py
@@ -5,7 +5,7 @@ from schemas.schemas import UserUpdateRequest, SignUpRequest, Results, ResultUse
 from services.services import UserServices
 from db.db_connection import get_db
 from databases import Database
-
+from services.auth_services import get_current_user
 user_routers = APIRouter()
 
 
@@ -15,14 +15,14 @@ def home():
 
 
 @user_routers.get("/users", response_model=Results)
-async def get_users(db: Database = Depends(get_db)) -> Results:
+async def get_users(db: Database = Depends(get_db), current_user=Depends(get_current_user)) -> Results:
     user_service = UserServices(db=db)
     users_list = await user_service.get_users()
     return users_list
 
 
 @user_routers.get("/user", response_model=ResultUser)
-async def get_user(user_id: int, db: Database = Depends(get_db)) -> ResultUser:
+async def get_user(user_id: int, db: Database = Depends(get_db), current_user=Depends(get_current_user)) -> ResultUser:
     user_service = UserServices(db=db)
     user_db = await user_service.get_user(user_id=user_id)
     if not user_db:
@@ -40,7 +40,10 @@ async def create_user(user: SignUpRequest, db: Database = Depends(get_db)) -> Re
 
 
 @user_routers.put("/user", response_model=ResultUser, status_code=status.HTTP_200_OK)
-async def update_user(user_id: int, user: UserUpdateRequest, db: Database = Depends(get_db)) -> ResultUser:
+async def update_user(user_id: int, user: UserUpdateRequest, current_user=Depends(get_current_user), db: Database = Depends(get_db)) -> ResultUser:
+    check_user_id = current_user.user_id
+    if user_id != check_user_id:
+        raise HTTPException(status_code=403, detail="It's not your account")
     user_service = UserServices(db=db)
     updated_user = await user_service.update_user(user_id=user_id, user=user)
     if not updated_user:
@@ -49,7 +52,10 @@ async def update_user(user_id: int, user: UserUpdateRequest, db: Database = Depe
 
 
 @user_routers.delete("/user", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_user(user_id: int, db: Database = Depends(get_db)):
+async def delete_user(user_id: int, db: Database = Depends(get_db), current_user=Depends(get_current_user)):
+    check_user_id = current_user.user_id
+    if user_id != check_user_id:
+        raise HTTPException(status_code=403, detail="It's not your account")
     user_service = UserServices(db=db)
     await user_service.delete_user(user_id=user_id)
